HW4_nba
Edited the NBA dataset to be efficient for the assingment (ommited some variables).
Applied boosting, bagging, random forests, and multiple linear regression on the same training data, and tested the models to compare which one fits the data best.

Still need to go over the linear again to make sure it's correct.
Still need to do logistic regression as well and compare it to the rest.